ccfri-5107 review changes
