+++ backend/src/routes/licence.js
@@ -3,15 +3,16 @@ const passport = require('passport');
 const router = express.Router();
 const auth = require('../components/auth');
 const isValidBackendToken = auth.isValidBackendToken();
-const { getLicenceDetails } = require('../components/licence');
-const { param, validationResult } = require('express-validator');
+const { getLicences } = require('../components/licence');
+const { query, validationResult } = require('express-validator');
 
 /**
- * Get licence details by licenceId.
+ * Get Licence details for a facility
+ *
  */
-router.get('/:licenceId', passport.authenticate('jwt', { session: false }), isValidBackendToken, [param('licenceId', 'URL param: [licenceId] is required').not().isEmpty()], (req, res) => {
+router.get('/', passport.authenticate('jwt', { session: false }), isValidBackendToken, query('facilityId', 'Query param: [facilityId] is required').notEmpty().isUUID(), (req, res) => {
   validationResult(req).throw();
-  return getLicenceDetails(req, res);
+  return getLicences(req, res);
 });
 
 module.exports = router;
