+++ frontend/src/services/licenceService.js
@@ -2,13 +2,13 @@ import ApiService from '@/common/apiService';
 import { ApiRoutes } from '@/utils/constants';
 
 export default {
-  async getLicenceDetails(licenceId) {
+  async getLicences(facilityId) {
     try {
-      if (!licenceId) return [];
-      const response = await ApiService.apiAxios.get(`${ApiRoutes.LICENCE_DETAILS}/${licenceId}`);
+      if (!facilityId) return [];
+      const response = await ApiService.apiAxios.get(`${ApiRoutes.LICENCES}?facilityId=${facilityId}`);
       return response?.data;
     } catch (error) {
-      console.log(`Failed to get licence detail(s) by licenceId - ${error}`);
+      console.log(`Failed to get the licences by facilityId - ${error}`);
       throw error;
     }
   },
