+++ frontend/src/utils/constants.js
@@ -57,7 +57,7 @@ export const ApiRoutes = Object.freeze({
   CANADA_POST: baseRoot + '/canadaPost',
   FUNDING_AGREEMENTS: baseRoot + '/fundingAgreements',
   ENROLMENT_REPORTS: baseRoot + '/enrolmentReports',
-  LICENCE_DETAILS: baseRoot + '/licences',
+  LICENCES: baseRoot + '/licences',
 });
 
 export const PAGE_TITLES = Object.freeze({
@@ -571,33 +571,12 @@ export const FUNDING_AGREEMENTS_STATUS = Object.freeze({
   EXPIRED: 'Expired',
 });
 
-export const DAYS_OF_WEEK = Object.freeze([
-  {
-    title: 'Monday',
-    value: 1,
-  },
-  {
-    title: 'Tuesday',
-    value: 2,
-  },
-  {
-    title: 'Wednesday',
-    value: 3,
-  },
-  {
-    title: 'Thursday',
-    value: 4,
-  },
-  {
-    title: 'Friday',
-    value: 5,
-  },
-  {
-    title: 'Saturday',
-    value: 6,
-  },
-  {
-    title: 'Sunday',
-    value: 7,
-  },
-]);
+export const DAYS_OF_WEEK = {
+  1: 'Monday',
+  2: 'Tuesday',
+  3: 'Wednesday',
+  4: 'Thursday',
+  5: 'Friday',
+  6: 'Saturday',
+  7: 'Sunday',
+};
